+++ src/daemon/mod.rs
@@ -181,7 +181,10 @@ pub(super) async fn start(
     )?)));
     let forest_car_db_dir = db_root_dir.join("car_db");
     load_all_forest_cars(&db, &forest_car_db_dir)?;
-    load_actor_bundles(&db).await?;
+
+    if config.client.load_actors {
+        load_actor_bundles(&db).await?;
+    }
 
     let mut services = JoinSet::new();